broke up token controls into separate files
